everything working with new hamlib_util file
